+++ identity_presentation/src/commonMain/kotlin/net/thechance/mena/identity/presentation/screen/forgetPassword/ForgetPasswordScreen.kt
@@ -21,7 +21,7 @@ import net.thechance.mena.designsystem.presentation.component.button.PrimaryButt
 import net.thechance.mena.designsystem.presentation.component.scaffold.Scaffold
 import net.thechance.mena.designsystem.presentation.theme.theme.Theme
 import net.thechance.mena.identity.presentation.base.BaseScreen
-import net.thechance.mena.identity.presentation.screen.countryPicker.CountryPicker
+import net.thechance.mena.identity.presentation.components.bottomSheet.countryPicker.CountryPickerBottomSheet
 import net.thechance.mena.identity.presentation.components.AuthAppBar
 import net.thechance.mena.identity.presentation.components.AuthScreenContainer
 import net.thechance.mena.identity.presentation.components.ErrorSnackBar
@@ -55,7 +55,7 @@ class ForgetPasswordScreen : BaseScreen<
         Scaffold(
             overlays = {
                 bottomSheet(isVisible = state.showCountryBottomSheet) { showBottomSheet ->
-                    CountryPicker(
+                    CountryPickerBottomSheet(
                         isVisible = showBottomSheet,
                         currentCountry = state.currentCountry,
                         onDismiss = listener::onDismissBottomSheet,
